requirements.txt
pandas>=1.3.0
snowflake-connector-python>=3.0.0
streamlit>=1.15.0
openai>=0.27.0
python-dotenv>=0.20.0